Rename TMMergeTracker to TranslationCounter
