+++ server/services/src/main/java/org/zanata/service/tm/merge/TMMergeResult.kt
@@ -2,20 +2,22 @@ package org.zanata.service.tm.merge
 
 import org.zanata.common.ContentState
 import org.zanata.common.ContentState.*
+import org.zanata.service.TranslationCounter
+import org.zanata.service.TextFlowCounter
 import java.util.*
 
 /**
  * @author Sean Flanigan <a href="mailto:sflaniga@redhat.com">sflaniga@redhat.com</a>
  */
-class TMMergeResult (private val bandDefs: Map<ContentState, List<IntRange>>): TMMergeTracker {
+class TMMergeResult (private val bandDefs: Map<ContentState, List<IntRange>>): TranslationCounter {
 
-    private val bandCounters = HashMap<Pair<ContentState, IntRange>, MutableCounter>()
+    private val bandCounters = HashMap<Pair<ContentState, IntRange>, MutableTextFlowCounter>()
 
     init {
         // for each defined ContentState/IntRange pair, we hold a Counter
         for((contentState, ranges) in bandDefs) {
             ranges.forEach { r ->
-                bandCounters[Pair(contentState, r)] = MutableCounter()
+                bandCounters[Pair(contentState, r)] = MutableTextFlowCounter()
             }
         }
     }
@@ -51,7 +53,7 @@ class TMMergeResult (private val bandDefs: Map<ContentState, List<IntRange>>): T
     /**
      * Returns a read-only counter for the specified (ContentState, IntRange)
      */
-    fun getCounter(state: ContentState, range: IntRange): Counter =
+    fun getCounter(state: ContentState, range: IntRange): TextFlowCounter =
         bandCounters[Pair(state, range)]!!
 
     override fun count(state: ContentState, score: Int, chars: Long, words: Long, messages: Long) {
@@ -67,26 +69,4 @@ class TMMergeResult (private val bandDefs: Map<ContentState, List<IntRange>>): T
     }
 }
 
-interface TMMergeTracker {
-    /**
-     * Count one or more copied messages
-     * @param state state of copied message
-     * @param score similarity score of the match (in source language)
-     * @param chars number of characters in the source language (Unicode code points)
-     * @param words number of words in the source language
-     * @param messages number of messages copied. Must be >= 1.
-     */
-    fun count(state: ContentState, score: Int, chars: Long, words: Long, messages: Long = 1)
-
-    object NOOP: TMMergeTracker {
-        override fun count(state: ContentState, score: Int, chars: Long, words: Long, messages: Long) {}
-    }
-}
-
-interface Counter {
-    val codePoints: Long
-    val words: Long
-    val messages: Long
-}
-
-private data class MutableCounter(override var codePoints: Long = 0, override var words: Long = 0, override var messages: Long = 0): Counter
+private data class MutableTextFlowCounter(override var codePoints: Long = 0, override var words: Long = 0, override var messages: Long = 0): TextFlowCounter
